Separate vehicleImages from images
